# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Argus API
  description: Multi-cluster Kubernetes admin dashboard with plugin system
  version: "1.0.0"
  contact:
    name: Argus Team
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/clusters:
    get:
      tags: [Clusters]
      summary: List all clusters
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cluster list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cluster"
    post:
      tags: [Clusters]
      summary: Add a new cluster
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, api_server_url, kubeconfig]
              properties:
                name:
                  type: string
                api_server_url:
                  type: string
                kubeconfig:
                  type: string
                  format: byte
      responses:
        "201":
          description: Cluster added
  /api/clusters/{id}:
    get:
      tags: [Clusters]
      summary: Get cluster details
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClusterId"
      responses:
        "200":
          description: Cluster details
    delete:
      tags: [Clusters]
      summary: Remove a cluster
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClusterId"
      responses:
        "204":
          description: Cluster removed
  /api/clusters/{id}/resources/{group}/{version}/{resource}:
    get:
      tags: [Resources]
      summary: List Kubernetes resources
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClusterId"
        - name: group
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: resource
          in: path
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Resource list
    post:
      tags: [Resources]
      summary: Create a Kubernetes resource
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ClusterId"
        - name: group
          in: path
          required: true
          schema:
            type: string
        - name: version
          in: path
          required: true
          schema:
            type: string
        - name: resource
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Resource created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ClusterId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Cluster:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        api_server_url:
          type: string
        status:
          type: string
          enum: [connected, disconnected, error, unreachable]
        labels:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
...